[PORT FROM MAIN] Fix crashes introduced when NUM_BUFFER_SLOTS increased to 64
BZ: 128174

Merge from MR1 main branch.

Change http://android.intel.com:8080/87673 increased the value of the
constant NUM_BUFFER_SLOTS from 32 to 64. The methods
BufferQueue:getReleasedBuffers and ConsumerBase::onBuffersReleased
were written assuming that NUM_BUFFER_SLOTS is 32 or less. If
NUM_BUFFER_SLOTS is greater than 32, buffers that are still in use are
sometimes freed, resulting in crashes and other catastrophic failures.
I changed those two methods so that they work correctly when
NUM_BUFFER_SLOTS is 64.

Category: aosp improvement
Domain: Graphics-SF
Origin: internal
Upstream-Candidate: no, proprietary

+++ include/gui/IGraphicBufferConsumer.h
@@ -153,7 +153,7 @@ public:
     // but have not yet been released by the consumer.
     //
     // This should be called from the onBuffersReleased() callback.
-    virtual status_t getReleasedBuffers(uint32_t* slotMask) = 0;
+    virtual status_t getReleasedBuffers(uint64_t* slotMask) = 0;
 
     // setDefaultBufferSize is used to set the size of buffers returned by
     // dequeueBuffer when a width and height of zero is requested.  Default
